+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpClassMethod.cs
@@ -112,6 +112,15 @@ public class CSharpClassMethod : CSharpMember<CSharpClassMethod>, IHasCSharpStat
         return this;
     }
 
+    public CSharpClassMethod InsertParameter(int index, string type, string name, Action<CSharpParameter> configure = null)
+    {
+        var param = new CSharpParameter(type, name);
+        Parameters.Insert(index, param);
+        configure?.Invoke(param);
+        return this;
+    }
+
+
     public CSharpClassMethod AddOptionalCancellationTokenParameter<T>(CSharpTemplateBase<T> template) =>
         AddParameter(
             $"{template.UseType("System.Threading.CancellationToken")}", "cancellationToken",